Fix issue with checking the request url when 2fa is disabled.
When logging in when 2fa is disabled in .env, the redirect could point to "/admin", but the middleware checks for 'admin/*', which fails the test and shows the 2fa requirement page. Going to any other page does not show the 2fa requirement page. This small change catches the '/admin' url.